+++ app/participants/controller.js
