+++ app/tournaments/controller.js
