+++ app/tournaments/details/controller.js
