+++ app/tournaments/index.js
@@ -1,4 +1,4 @@
 module.exports = app => {
-    require('./ctrl.js')(app);
+    require('./controller.js')(app);
     require('./details')(app);
 };
